% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{grpc_status_codes}
\alias{grpc_status_codes}
\title{gRPC status codes}
\format{
An object of class \code{integer} of length 17.
}
\usage{
grpc_status_codes
}
\description{
Named integer vector mapping gRPC status names to their wire codes.
}
\keyword{datasets}
