% tinyrox says don't edit this manually, but it can't stop you!
\name{grpc_version}
\alias{grpc_version}
\title{Version of the linked gRPC C++ library}
\usage{
grpc_version()
}
\value{
A character string, e.g. \code{"1.51.1"}.
}
\description{
Returns the version string of the system gRPC C++ library this package
was built against.
}
\examples{
grpc_version()
}
